pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "risk-modeling-q1"
version = "0.1.0"
description = "Risk modeling pipeline — Q1"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "lightgbm>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "jupyter>=1.0",
    "ipykernel>=6.25",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
